NEW: Add phpstan to travis testing, via Symbiote module.
Note that this is a bit of a WIP because it’s grabbing Sam’s
unmerged PR for silverstripe-phpstan to get it working with the latest
phpstan version. However, that knocks out a bunch of false-positives.

+++ src/Core/Convert.php
@@ -314,8 +314,7 @@ class Convert
      * Convert a XML string to a PHP array recursively. Do not
      * call this function directly, Please use {@link Convert::xml2array()}
      *
-     * @param SimpleXMLElement
-     *
+     * @param SimpleXMLElement $xml
      * @return mixed
      */
     protected static function recursiveXMLToArray($xml)
@@ -452,7 +451,7 @@ class Convert
      * Please only encode the values, not the whole url, e.g.
      * "mailto:test@test.com?subject=" . Convert::raw2mailto($subject)
      *
-     * @param $data string
+     * @param string $data
      * @return string
      * @see http://www.ietf.org/rfc/rfc1738.txt
      */
@@ -469,7 +468,7 @@ class Convert
      * Convert a string (normally a title) to a string suitable for using in
      * urls and other html attributes. Uses {@link URLSegmentFilter}.
      *
-     * @param string
+     * @param string $title
      * @return string
      */
     public static function raw2url($title)
@@ -528,7 +527,7 @@ class Convert
      * - IDField => idField
      * - iDField => iDField
      *
-     * @param $str
+     * @param string $str
      * @return string
      */
     public static function upperCamelToLowerCamel($str)
@@ -574,7 +573,7 @@ class Convert
         // Remove  non-unit characters from the size
         $unit = preg_replace('/[^bkmgtpezy]/i', '', $memString);
         // Remove non-numeric characters from the size
-        $size = preg_replace('/[^0-9\.\-]/', '', $memString);
+        $size = (float)preg_replace('/[^0-9\.\-]/', '', $memString);
 
         if ($unit) {
             // Find the position of the unit in the ordered string which is the power
